Begin refactoring FileUtilities.cpp
