+++ src/Metropolis/Utilities/FileUtilities.cpp
@@ -489,10 +489,9 @@ bool buildBoxData(Environment* enviro, vector<Molecule>& molecVec, Box* box, SBS
   return true;
 }
 
-bool generatefccBox(Box* box)
-{
-	if (box->environment == NULL || box->molecules == NULL)
-	{
+bool generatefccBox(Box* box) {
+	// Return an error if the box is missing an environment or molecules.
+	if (box->environment == NULL || box->molecules == NULL) {
 		std::cerr << "Error: generatefccBox(): Box environment or molecules array NULL" << std::endl;
 		return false;
 	}
@@ -507,70 +506,59 @@ bool generatefccBox(Box* box)
 
 	//Check if numOfMolecules is a non-fcc number of molecules
 	//and increase the number of cells if necessary
-	while((4 * cells * cells * cells) < enviro->numOfMolecules)
-    {
+	while((4 * cells * cells * cells) < enviro->numOfMolecules) {
 		cells++;
-    }
+  }
 
 	//Determine length of unit cell
-	cellL = enviro->x/ (Real) cells;
+	cellL = enviro->x / (Real) cells;
 	halfcellL = 0.5 * cellL;
 
 	//Construct the unit cell
-	for (int j = 0; j < molecules[0].numOfAtoms; j++)
-    {
-    	molecules[0].atoms[j].x += 0.0;
-        molecules[0].atoms[j].y += 0.0;
-        molecules[0].atoms[j].z += 0.0;
+	for (int j = 0; j < molecules[0].numOfAtoms; j++) {
+    molecules[0].atoms[j].x += 0.0;
+    molecules[0].atoms[j].y += 0.0;
+    molecules[0].atoms[j].z += 0.0;
 	}
 
-	for (int j = 0; j < molecules[1].numOfAtoms; j++)
-    {
-    	molecules[1].atoms[j].x += halfcellL;
-        molecules[1].atoms[j].y += halfcellL;
-        molecules[1].atoms[j].z += 0.0;
-    }
+	for (int j = 0; j < molecules[1].numOfAtoms; j++) {
+    molecules[1].atoms[j].x += halfcellL;
+    molecules[1].atoms[j].y += halfcellL;
+    molecules[1].atoms[j].z += 0.0;
+  }
 
-    for (int j = 0; j < molecules[2].numOfAtoms; j++)
-    {
-        molecules[2].atoms[j].x += 0.0;
-        molecules[2].atoms[j].y += halfcellL;
-        molecules[2].atoms[j].z += halfcellL;
-    }
+  for (int j = 0; j < molecules[2].numOfAtoms; j++) {
+  	molecules[2].atoms[j].x += 0.0;
+    molecules[2].atoms[j].y += halfcellL;
+    molecules[2].atoms[j].z += halfcellL;
+  }
 
-    for (int j = 0; j < molecules[3].numOfAtoms; j++)
-    {
-        molecules[3].atoms[j].x += halfcellL;
-        molecules[3].atoms[j].y += 0.0;
-        molecules[3].atoms[j].z += halfcellL;
-    }
+  for (int j = 0; j < molecules[3].numOfAtoms; j++) {
+    molecules[3].atoms[j].x += halfcellL;
+    molecules[3].atoms[j].y += 0.0;
+    molecules[3].atoms[j].z += halfcellL;
+  }
 
 	//Init all other molecules to initial coordinates
 	//Build the lattice from the unit cell by repeatedly translating
 	//the four vectors of the unit cell through a distance cellL in
 	//the x, y, and z directions
-	for(int i = 4; i < enviro->numOfMolecules; i++)
-    {
-		for (int j = 0; j < molecules[i].numOfAtoms; j++)
-        {
+	for(int i = 4; i < enviro->numOfMolecules; i++) {
+		for (int j = 0; j < molecules[i].numOfAtoms; j++) {
 			molecules[i].atoms[j].x += 0.0;
-    		molecules[i].atoms[j].y += 0.0;
-   	 		molecules[i].atoms[j].z += 0.0;
-   	 	}
+    	molecules[i].atoms[j].y += 0.0;
+   	 	molecules[i].atoms[j].z += 0.0;
+   	}
 	}
 
 	int offset = 0;
-	for(int z = 1; z <= cells; z++)
-		for(int y = 1; y <= cells; y++)
-			for(int x = 1; x <= cells; x++)
-            {
-				for(int a = 0; a < 4; a++)
-                {
+	for (int z = 1; z <= cells; z++) {
+		for (int y = 1; y <= cells; y++) {
+			for (int x = 1; x <= cells; x++) {
+				for (int a = 0; a < 4; a++) {
 					int i = a + offset;
-					if(i < enviro->numOfMolecules)
-                    {
-						for (int j = 0; j < molecules[i].numOfAtoms; j++)
-                        {
+					if (i < enviro->numOfMolecules) {
+						for (int j = 0; j < molecules[i].numOfAtoms; j++) {
 							molecules[i].atoms[j].x = molecules[a].atoms[j].x + cellL * (x-1);
 							molecules[i].atoms[j].y = molecules[a].atoms[j].y + cellL * (y-1);
 							molecules[i].atoms[j].z = molecules[a].atoms[j].z + cellL * (z-1);
@@ -579,12 +567,12 @@ bool generatefccBox(Box* box)
 				}
 				offset += 4;
 			}
+		}
+	}
 
 	//Shift center of box to the origin
-	for(int i = 0; i < enviro->numOfMolecules; i++)
-    {
-		for (int j = 0; j < molecules[i].numOfAtoms; j++)
-        {
+	for (int i = 0; i < enviro->numOfMolecules; i++) {
+		for (int j = 0; j < molecules[i].numOfAtoms; j++) {
 			molecules[i].atoms[j].x -= halfcellL;
 			molecules[i].atoms[j].y -= halfcellL;
 			molecules[i].atoms[j].z -= halfcellL;
@@ -594,1163 +582,6 @@ bool generatefccBox(Box* box)
 	return true;
 }
 
-ConfigScanner::ConfigScanner()
-{
-	enviro = Environment();
-	numOfSteps = 0;
-	useStatefileSetup = false;
-	useZMatrixSetup = false;
-}
-
-
-bool ConfigScanner::readInConfig(string configpath)
-{
-	isSafeToContinue = true;
-	enviro = Environment();
-	numOfSteps = 0;
-
-	configPath = configpath;
-	if (configPath.empty())
-	{
-		std::cerr << "Configuration File path is empty" << std::endl;
-		return false;
-	}
-
-    ifstream configscanner(configPath.c_str());
-    if (!configscanner.is_open())
-    {
-        std::cerr << "Unable to open configuration file (" << configPath << ")" << std::endl;
-        return false;
-    }
-    else
-    {
-        string line;
-        int currentLine = 1;
-        while (configscanner.good())
-        {
-            getline(configscanner,line);
-			trim(line);
-            //assigns attributes based on line number
-            switch(currentLine)
-            {
-                case 2:
-					if(line.length() > 0)
-					{
-						enviro.x = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Error: Config File: Missing environment X value.");
-						return false;
-					}
-                    break;
-                case 3:
-					if(line.length() > 0)
-					{
-						enviro.y = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Error: Config File: Missing environment Y value.");
-						return false;
-					}
-                    break;
-                case 4:
-					if(line.length() > 0)
-					{
-						enviro.z = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment z value.");
-						return false;
-					}
-                    break;
-                case 6:
-					if(line.length() > 0)
-					{
-						enviro.temp = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment temperature value.");
-						return false;
-					}
-                    break;
-                case 8:
-					if(line.length() > 0)
-					{
-						enviro.maxTranslation = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment max translation value.");
-						return false;
-					}
-                    break;
-                case 10:
-					if(line.length() > 0)
-					{
-						numOfSteps = atoi(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing number of steps value.");
-						return false;
-					}
-                    break;
-                case 12:
-					if(line.length() > 0)
-					{
-						enviro.numOfMolecules = atoi(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing number of molecules value.");
-						return false;
-					}
-                    break;
-                case 14:
-					if(line.length() > 0)
-					{
-						oplsuaparPath = line;
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing oplsuapar path value.");
-						return false;
-					}
-                    break;
-                case 16:
-					if(line.length() > 0)
-					{
-						zmatrixPath = line;
-						useZMatrixSetup = true;
-						useStatefileSetup = true;
-					}
-					else
-					{
-						throwScanError("INFO: Configuration file not well formed. Missing z-matrix path value. Attempting to rely on prior state information...");
-						isSafeToContinue = false; //now that it is false, check if there is a state file
-						useZMatrixSetup = false;
-						useStatefileSetup = true;
-					}
-					break;
-                case 18:
-                    if(line.length() > 0)
-					{
-                        statePath = line;
-                        useZMatrixSetup = false; //we will go ahead and use the statefile, since the path was provided
-                        useStatefileSetup = true;
-                    }
-                    else if (isSafeToContinue == false)
-					{
-						throwScanError("Configuration file not well formed. Missing value pointing to prior state file path. Cannot safely continue with program execution.");
-						useStatefileSetup = false; //we can't even hope to find the state file, so don't use it
-						useZMatrixSetup = false;
-						return false; //preferable to simply exiting, as we want to give the program a chance to do...whatever?
-					}
-					else
-					{
-						throwScanError("INFO: Value pointing to prior state file path not found in main config file. Environment setup defaulting to clean simulation.");
-						useZMatrixSetup = true; //revert to using the Zmatrix file for all setup
-						useStatefileSetup = false;
-						isSafeToContinue = true;
-					}
-					break;
-                case 20:
-                    if(line.length() > 0){
-                        stateOutputPath = line;
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing state file output path value.");
-						return false;
-					}
-                    break;
-                case 22:
-                    if(line.length() > 0){
-                        pdbOutputPath = line;
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing PDB output path value.");
-						return false;
-					}
-                    break;
-                case 24:
-					if(line.length() > 0)
-					{
-						enviro.cutoff = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment cutoff value.");
-						return false;
-					}
-                    break;
-                case 26:
-					if(line.length() > 0)
-					{
-						enviro.maxRotation = atof(line.c_str());
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment max rotation value.");
-						return false;
-					}
-                    break;
-                case 28:
-                    if(line.length() > 0){
-						enviro.randomseed=atoi(line.c_str());
-                    }
-                    else
-					{
-						enviro.randomseed = (unsigned int) time(NULL);
-						//throwScanError("Configuration file not well formed. Missing random seed value.");
-						//return false;
-					}
-                	break;
-                case 30:
-                    if(line.length() > 0){
-						// Convert to a zero-based index
-						parsePrimaryIndexDefinitions(line);
-                    }
-					else
-					{
-						throwScanError("Configuration file not well formed. Missing environment primary atom index value.");
-						return false;
-					}
-                	break;
-            }
-
-			currentLine++;
-        }
-    }
-
-    configscanner.close();
-
-    return true;
-}
-
-void ConfigScanner::parsePrimaryIndexDefinitions(string definitions)
-{
-    *enviro.primaryAtomIndexConfigLine = definitions;
-    char *indexVector;
-    char *charLine = (char *) malloc(sizeof(char) * (definitions.size()+1));
-    strcpy(charLine, definitions.c_str());
-    indexVector = strtok(charLine, ",");
-
-    for (int i = 0; indexVector ; i++)
-    {
-	std::string sIndexVector = indexVector;
-	enviro.primaryAtomIndexDefinitions++;
-	char* c;
-	if ((c=strchr(indexVector, '['))!=NULL)
-	{
-	    indexVector = (indexVector + 1);
-	    while ((c=strchr(indexVector, ']'))==NULL)
-	    {
-		int currentPrimaryIndex = atoi(indexVector);
-		(*(enviro.primaryAtomIndexArray)).push_back(new std::vector<int>);
-		(*(*(enviro.primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-		indexVector = strtok(NULL, ",");
-	    }
-	    *c = '\0';
-	    int currentPrimaryIndex = atoi(indexVector);
-	    (*(enviro.primaryAtomIndexArray)).push_back(new std::vector<int>);
-	    (*(*(enviro.primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-	    indexVector =strtok(NULL, ",");
-	    continue;
-	}
-
-        int currentPrimaryIndex = atoi(indexVector);
-	(*(enviro.primaryAtomIndexArray)).push_back(new std::vector<int>);
-	(*(*(enviro.primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-	indexVector = strtok(NULL, ",");
-    }
-    free (charLine);
-}
-
-void ConfigScanner::throwScanError(string message)
-{
-
-	std::cerr << message << std::endl;
-}
-
-Environment* ConfigScanner::getEnviro()
-{
-    return &enviro;
-}
-
-string ConfigScanner::getConfigPath()
-{
-    return configPath;
-}
-
-long ConfigScanner::getSteps()
-{
-    return numOfSteps;
-}
-
-string ConfigScanner::getOplsusaparPath()
-{
-    return oplsuaparPath;
-}
-
-string ConfigScanner::getZmatrixPath()
-{
-    return zmatrixPath;
-}
-
-string ConfigScanner::getStatePath()
-{
-    return statePath;
-}
-
-string ConfigScanner::getStateOutputPath()
-{
-    return stateOutputPath;
-}
-
-string ConfigScanner::getPdbOutputPath()
-{
-    return pdbOutputPath;
-}
-
-// ============================================================================
-// ============================= OPLS SCANNER =================================
-// ============================================================================
-
-
-OplsScanner::OplsScanner()
-{
-}
-
-OplsScanner::~OplsScanner()
-{
-    oplsTable.clear();
-    fourierTable.clear();
-}
-
-bool OplsScanner::readInOpls(string filename)
-{
-	fileName = filename;
-	if (filename.empty())
-	{
-		std::cerr << "Error: readInOpls(): empty filename given" << std::endl;
-		return false;
-	}
-
-    int numOfLines=0;
-    ifstream oplsScanner(filename.c_str());
-    if (!oplsScanner.is_open())
-    {
-    	std::cerr << "Error: readInOpls(): could not open file (" << filename << ")" << std::endl;
-        return false;
-    }
-    else
-    {
-        string line;
-        while (oplsScanner.good())
-        {
-            numOfLines++;
-            getline(oplsScanner,line);
-
-            //check if it is a commented line,
-            //or if it is a title line
-            try
-            {
-                if(line.at(0) != '#' && numOfLines > 1)
-		{
-                    addLineToTable(line,numOfLines);
-		}
-            }
-            catch (std::out_of_range& e)
-            {
-            	// Eat the exception and continue...why aren't we failing?
-            }
-        }
-        oplsScanner.close();
-        logErrors();
-    }
-
-    return true;
-}
-
-void OplsScanner::addLineToTable(string line, int numOfLines)
-{
-    string hashNum;
-    int secCol;
-    Real charge,sigma,epsilon;
-    string name, extra;
-    stringstream ss(line);
-
-    //check to see what format it is opls, V value, or neither
-    int format = checkFormat(line);
-
-    if(format == 1)
-    {
-        ss >> hashNum >> secCol >> name >> charge >> sigma >> epsilon;
-	   char *atomType = (char *)name.c_str();
-
-        Atom temp = createAtom(0, -1, -1, -1, sigma, epsilon, charge, name);
-
-        pair<map<string,Atom>::iterator,bool> ret;
-        ret = oplsTable.insert( pair<string,Atom>(hashNum,temp) );
-
-        if (ret.second==false)
-        {
-            errHashes.push_back(hashNum);
-        }
-    }
-    else if(format == 2)
-    {
-        Real v0,v1,v2,v3;
-        ss >> hashNum >> v0 >> v1 >> v2 >> v3 ;
-        Fourier vValues = {v0,v1,v2,v3};
-        pair<map<string,Fourier>::iterator,bool> ret2;
-        ret2 = fourierTable.insert( pair<string,Fourier>(hashNum,vValues) );
-
-        if (ret2.second==false)
-        {
-            errHashesFourier.push_back(hashNum);
-        }
-    }
-    else
-    {
-	     errLines.push_back(numOfLines);
-    }
-}
-
-int OplsScanner::checkFormat(string line)
-{
-    int hashNum, secCol;
-    Real charge,sigma,epsilon;
-    string name, extra;
-    stringstream iss(line);
-
-    Real v1,v2,v3,v4;
-    stringstream issw(line);
-
-    //see if format is the V values for the diherdral format
-    if((issw >> hashNum >> v1 >> v2 >> v3 >> v4) )
-    {
-        return 2;
-    }
-    //else see if format is normal opls format
-    else if((iss >> hashNum >> secCol >> name >> charge >> sigma >> epsilon ))
-    {
-        return 1;
-    }
-    //if neither return -1
-    else
-    {
-        return -1;
-    }
-}
-
-void OplsScanner::logErrors()
-{
-    stringstream output;
-    // See if there were any errors
-    if(errLines.empty() || errHashes.empty()|| errHashesFourier.empty())
-    {
-	     //Errors in the format
-		  output<<"Errors found in the OPLS file: "<< fileName<<endl;
-        if(!errLines.empty())
-        {
-		      output << "Found Errors in the Format of the following Lines: " << endl;
-				for(int a=0; a<errLines.size(); a++)
-                {
-				    if(a%10==0 && a!=0) //ten per line
-                    {
-					     output << endl;
-                    }
-				    output << errLines[a]<< " ";
-				}
-				output << endl<< endl;
-		  }
-		  if(!errHashes.empty())
-          {
-		      output << "Error - The following OPLS values existed more than once: " << endl;
-				for(int a=0; a<errHashes.size(); a++)
-                {
-				    if(a%10==0 && a!=0) //ten per line
-                    {
-					     output << endl;
-                    }
-				    output << errHashes[a]<< " ";
-				}
-				output << endl<< endl;
-		  }
-		  if(!errHashesFourier.empty())
-          {
-		      output << "Error - The following Fourier Coefficent values existed more than once: " << endl;
-				for(int a=0; a<errHashesFourier.size(); a++)
-                {
-				    if(a%10==0 && a!=0) //ten per line
-                    {
-					     output << endl;
-                    }
-				    output << errHashesFourier[a]<< " ";
-				}
-				output << endl<< endl;
-		  }
-		  writeToLog(output,OPLS);
-	}
-}
-
-Atom OplsScanner::getAtom(string hashNum)
-{
-    if(oplsTable.count(hashNum)>0 )
-    {
-        return oplsTable[hashNum];
-	}
-	else
-    {
-	    cerr << "Index does not exist: "<< hashNum <<endl;
-		return createAtom(0, -1, -1, -1, -1, -1, -1, NULL);
-	}
-}
-
-Real OplsScanner::getSigma(string hashNum)
-{
-    if(oplsTable.count(hashNum)>0 )
-    {
-        Atom temp = oplsTable[hashNum];
-        return temp.sigma;
-    }
-    else
-    {
-        cerr << "Index does not exist: "<< hashNum <<endl;
-        return -1;
-    }
-}
-
-Real OplsScanner::getEpsilon(string hashNum)
-{
-    if(oplsTable.count(hashNum)>0 )
-    {
-        Atom temp = oplsTable[hashNum];
-        return temp.epsilon;
-    }
-    else
-    {
-        cerr << "Index does not exist: "<< hashNum <<endl;
-        return -1;
-    }
-}
-
-Real OplsScanner::getCharge(string hashNum)
-{
-    if(oplsTable.count(hashNum)>0 )
-    {
-        Atom temp = oplsTable[hashNum];
-        return temp.charge;
-    }
-    else
-    {
-        cerr << "Index does not exist: "<< hashNum <<endl;
-        return -1;
-    }
-}
-
-Fourier OplsScanner::getFourier(string hashNum)
-{
-    if(fourierTable.count(hashNum)>0 )
-    {
-        Fourier temp = fourierTable[hashNum];
-        return temp;
-    }
-    else
-    {
-        cerr << "Index does not exist: "<< hashNum <<endl;
-        Fourier temp ={-1,-1,-1,-1};
-        return temp;
-    }
-}
-
-// ============================================================================
-// ======================== Z-Matrix Scanner ==================================
-// ============================================================================
-
-ZmatrixScanner::ZmatrixScanner()
-{
-    oplsScanner = NULL;
-    startNewMolecule = false;
-    previousFormat = 0;
-}
-
-ZmatrixScanner::~ZmatrixScanner()
-{
-}
-
-bool ZmatrixScanner::readInZmatrix(string filename, OplsScanner* scanner)
-{
-	fileName = filename;
-	oplsScanner = scanner;
-	startNewMolecule = false;
-
-	if (fileName.empty())
-	{
-		std::cerr << "Error: readInZmatrix(): Given filename is NULL" << std::endl;
-		return false;
-	}
-
-    stringstream output;
-    int numOfLines=0;
-
-    ifstream zmatrixScanner(fileName.c_str());
-    if( !zmatrixScanner.is_open() )
-    {
-    	std::cerr << "Error: Unable to open Z-Matrix file (" << fileName << ")" << std::endl;
-        return false;
-    }
-    else
-    {
-        string line;
-	    int moleculeNum = -1;
-
-        while( zmatrixScanner.good() )
-        {
-            numOfLines++;
-            getline(zmatrixScanner,line);
-
-            Molecule workingMolecule;
-	        workingMolecule.type = moleculeNum;
-            try
-            {
-                if(line.at(0) != '#' && numOfLines > 1)
-                {
-                    parseLine(line,numOfLines);
-                }
-            }
-            catch(std::out_of_range& e){}
-
-            if (startNewMolecule)
-            {
-                Atom* atomArray;
-                Bond* bondArray;
-                Angle* angleArray;
-                Dihedral* dihedralArray;
-                moleculeNum++;
-                workingMolecule.type = moleculeNum;
-
-		        double atomVectorSize = atomVector.size();
-		        int sizeOfAtom = sizeof(Atom);
-
-		        atomArray = (Atom*) malloc(sizeof(Atom) * atomVector.size());
-                bondArray = (Bond*) malloc(sizeof(Bond) * bondVector.size());
-                angleArray = (Angle*) malloc(sizeof(Angle) * angleVector.size());
-                dihedralArray = (Dihedral*) malloc(sizeof(Dihedral) * dihedralVector.size());
-
-                for (int i = 0; i < atomVector.size(); i++)
-                {
-                    atomArray[i] = atomVector[i];
-                }
-                for (int i = 0; i < bondVector.size(); i++)
-                {
-                    bondArray[i] = bondVector[i];
-                }
-                for (int i = 0; i < angleVector.size(); i++)
-                {
-                    angleArray[i] = angleVector[i];
-                }
-                for (int i = 0; i < dihedralVector.size(); i++)
-                {
-                    dihedralArray[i] = dihedralVector[i];
-                }
-
-                moleculePattern.push_back(createMolecule(-1, moleculeNum, atomArray, angleArray, bondArray, dihedralArray,
-                     atomVector.size(), angleVector.size(), bondVector.size(), dihedralVector.size()));
-
-                atomVector.clear();
-                bondVector.clear();
-                angleVector.clear();
-                dihedralVector.clear();
-
-                startNewMolecule = false;
-            }
-        }
-    }
-
-    zmatrixScanner.close();
-
-    return true;
-}
-
-void ZmatrixScanner::parseLine(string line, int numOfLines)
-{
-
-    string atomID, atomType, oplsA, oplsB, bondWith, bondDistance, angleWith, angleMeasure, dihedralWith, dihedralMeasure;
-
-    stringstream ss;
-
-    //check if line contains correct format
-    int format = checkFormat(line);
-
-    if(format == 1)
-    {
-        //read in strings in columns and store the data in temporary variables
-        ss << line;
-        ss >> atomID >> atomType >> oplsA >> oplsB >> bondWith >> bondDistance >> angleWith >> angleMeasure >> dihedralWith >> dihedralMeasure;
-
-	    //setup structures for permanent encapsulation
-        Atom lineAtom;
-        Bond lineBond;
-        Angle lineAngle;
-        Dihedral lineDihedral;
-
-        if (oplsA.compare("-1") != 0)
-        {
-            lineAtom = oplsScanner->getAtom(oplsA);
-            lineAtom.id = atoi(atomID.c_str());
-            lineAtom.x = 0;
-            lineAtom.y = 0;
-            lineAtom.z = 0;
-	        *lineAtom.name = atomType;
-        }
-        else
-        {
-            lineAtom = createAtom(atoi(atomID.c_str()), -1, -1, -1, -1, -1, -1, atomType);
-        }
-
-	    atomVector.push_back(lineAtom);
-
-        if (bondWith.compare("0") != 0)
-        {
-            lineBond.atom1 = lineAtom.id;
-            lineBond.atom2 = atoi(bondWith.c_str());
-            lineBond.distance = atof(bondDistance.c_str());
-            lineBond.variable = false;
-            bondVector.push_back(lineBond);
-        }
-
-        if (angleWith.compare("0") != 0)
-        {
-            lineAngle.atom1 = lineAtom.id;
-            lineAngle.atom2 = atoi(angleWith.c_str());
-            lineAngle.value = atof(angleMeasure.c_str());
-            lineAngle.variable = false;
-            angleVector.push_back(lineAngle);
-        }
-
-        if (dihedralWith.compare("0") != 0)
-        {
-            lineDihedral.atom1 = lineAtom.id;
-            lineDihedral.atom2 = atoi(dihedralWith.c_str());
-            lineDihedral.value = atof(dihedralMeasure.c_str());
-            lineDihedral.variable = false;
-            dihedralVector.push_back(lineDihedral);
-        }
-    }
-    else if(format == 2)
-    {
-        startNewMolecule = true;
-    }
-    else if(format == 3)
-    {
-        startNewMolecule = true;
-    }
-
-    if (previousFormat >= 3 && format == -1)
-    {
-        handleZAdditions(line, previousFormat);
-    }
-
-    previousFormat = format;
-}
-
-int ZmatrixScanner::checkFormat(string line)
-{
-    int format =-1;
-    stringstream iss(line);
-    stringstream iss2(line);
-    string atomType, someLine;
-    int atomID, oplsA, oplsB, bondWith, angleWith,dihedralWith,extra;
-    Real bondDistance, angleMeasure, dihedralMeasure;
-
-    // check if it is the normal 11 line format
-    if (iss >> atomID >> atomType >> oplsA >> oplsB >> bondWith >> bondDistance >>
-            angleWith >> angleMeasure >> dihedralWith >> dihedralMeasure >> extra)
-    {
-        format = 1;
-    }
-    else
-    {
-        someLine = line;
-        if(someLine.find("TERZ")!=string::npos)
-        {
-            format = 2;
-		}
-        else if(someLine.find("Geometry Variations")!=string::npos)
-        {
-            format = 3;
-        }
-        else if(someLine.find("Variable Bonds")!=string::npos)
-        {
-            format = 4;
-        }
-        else if(someLine.find("Additional Bonds")!=string::npos)
-        {
-            format = 5;
-        }
-        else if(someLine.find("Harmonic Constraints")!=string::npos)
-        {
-            format = 6;
-        }
-        else if(someLine.find("Variable Bond Angles")!=string::npos)
-        {
-            format = 7;
-        }
-        else if(someLine.find("Additional Bond Angles")!=string::npos)
-        {
-            format = 8;
-        }
-        else if(someLine.find("Variable Dihedrals")!=string::npos)
-        {
-            format = 9;
-        }
-        else if(someLine.find("Additional Dihedrals")!=string::npos)
-        {
-            format = 10;
-        }
-        else if(someLine.find("Domain Definitions")!=string::npos)
-        {
-            format = 11;
-        }
-        else if(someLine.find("Final blank line")!=string::npos)
-        {
-            format = -2;
-        }
-    }
-
-    return format;
-}
-
-void ZmatrixScanner::handleZAdditions(string line, int cmdFormat)
-{
-    vector<int> atomIds;
-    int id;
-    stringstream tss(line.substr(0,15) );
-
-    if(line.find("AUTO")!=string::npos)
-    {
-	     //Do stuff for AUTO
-	     //but what is AUTO-related stuff?
-    }
-    else
-    {
-        while(tss >> id)
-        {
-            atomIds.push_back(id);
-            if(tss.peek()=='-'||tss.peek()==','||tss.peek()==' ')
-            {
-                tss.ignore();
-            }
-        }
-
-        int start, end=0;
-        if( atomIds.size()== 1)
-        {
-            start = atomIds[0];
-            end = atomIds[0];
-        }
-        else if(atomIds.size() == 2)
-        {
-            start = atomIds[0]; end = atomIds[1];
-        }
-
-        switch(cmdFormat)
-        {
-            case 3:
-                // Geometry Variations follow
-            break;
-            case 4:
-                // Variable Bonds follow
-                for(int i=0; i< moleculePattern[0].numOfBonds; i++)
-                {
-                    if(  moleculePattern[0].bonds[i].atom1 >= start &&  moleculePattern[0].bonds[i].atom1 <= end)
-                    {
-                        moleculePattern[0].bonds[i].variable = true;
-                    }
-                }
-                break;
-            case 5:
-                //  Additional Bonds follow
-                break;
-            case 6:
-                // Harmonic Constraints follow
-                break;
-            case 7:
-                //  Variable Bond Angles follow
-                for(int i=0; i<  moleculePattern[0].numOfAngles; i++)
-                {
-                    if(  moleculePattern[0].angles[i].atom1 >= start && moleculePattern[0].angles[i].atom1 <= end)
-                    {
-                        moleculePattern[0].angles[i].variable = true;
-                    }
-                }
-                break;
-            case 8:
-                // Additional Bond Angles follow
-                break;
-            case 9:
-                // Variable Dihedrals follow
-                for(int i=0; i< moleculePattern[0].numOfDihedrals; i++)
-                {
-                    if(  moleculePattern[0].dihedrals[i].atom1 >= start &&  moleculePattern[0].dihedrals[i].atom1 <= end )
-                    {
-                        moleculePattern[0].dihedrals[i].variable = true;
-                    }
-                }
-                break;
-            case 10:
-                //  Domain Definitions follow
-                break;
-            default:
-                //Do nothing
-                break;
-        }
-    }
-}
-
-vector<Hop> ZmatrixScanner::calculateHops(Molecule molec)
-{
-    vector<Hop> newHops;
-    int **graph;
-    int size = molec.numOfAtoms;
-	int startId = molec.atoms[0].id;
-
-    buildAdjacencyMatrix(graph,molec);
-
-    for(int atom1=0; atom1<size; atom1++)
-    {
-        for(int atom2=atom1+1; atom2<size; atom2++)
-        {
-            int distance = findHopDistance(atom1,atom2,size,graph);
-            if(distance >=3)
-            {
-				Hop tempHop = Hop(atom1+startId,atom2+startId,distance); //+startId because atoms may not start at 1
-                newHops.push_back(tempHop);
-            }
-        }
-    }
-
-    return newHops;
-}
-
-bool ZmatrixScanner::contains(vector<int> &vect, int item)
-{
-    for(int i=0; i<vect.size(); i++)
-    {
-        if(vect[i]==item)
-        {
-            return true;
-        }
-    }
-
-    return false;
-}
-
-
-int ZmatrixScanner::findHopDistance(int atom1,int atom2,int size, int **graph)
-{
-    map<int,int> distance;
-    queue<int> Queue;
-    vector<int> checked;
-    vector<int> bonds;
-
-
-    Queue.push(atom1);
-    checked.push_back(atom1);
-    distance.insert( pair<int,int>(atom1,0) );
-
-    while(!Queue.empty())
-    {
-        int target = Queue.front();
-        Queue.pop();
-        if(target == atom2)
-        {
-            return distance[target];
-        }
-
-        bonds.clear();
-        for(int col=0;col<size;col++)
-        {
-            if( graph[target][col]==1 )
-            {
-                bonds.push_back(col);
-            }
-        }
-
-        for(int x=0; x<bonds.size();x++)
-        {
-            int currentBond = bonds[x];
-            if(!contains(checked,currentBond) )
-            {
-                checked.push_back(currentBond);
-                int newDistance = distance[target]+1;
-                distance.insert(pair<int,int>(currentBond, newDistance));
-                Queue.push(currentBond);
-            }
-        }
-    }
-
-    //Needs a return value
-    return -1; //Temp fill
-}
-
-void ZmatrixScanner::buildAdjacencyMatrix(int **&graph, Molecule molec)
-{
-    int size = molec.numOfAtoms;
-	int startId = molec.atoms[0].id;
-	int lastId = startId + molec.numOfAtoms -1;
-
-    graph =  new int*[size]; //create colums
-    for(int i=0; i<size; i++) //create rows
-    {
-        graph[i]=new int[size];
-    }
-
-    //fill with zero
-    for(int c=0; c<size; c++)
-    {
-        for(int r=0; r<size; r++)
-        {
-            graph[c][r]=0;
-        }
-    }
-
-    //fill with adjacent array with bonds
-    for(int x=0; x<molec.numOfBonds; x++)
-    {
-        Bond bond = molec.bonds[x];
-		  //make sure the bond is intermolecular
-		if( (bond.atom1 >= startId && bond.atom1 <= lastId) &&
-		   (bond.atom2 >= startId && bond.atom2 <= lastId) )
-        {
-			graph[bond.atom1-startId][bond.atom2-startId]=1;
-            graph[bond.atom2-startId][bond.atom1-startId]=1;
-		}
-    }
-}
-
-vector<Molecule> ZmatrixScanner::buildMolecule(int startingID)
-{
-	int numOfMolec = moleculePattern.size();
-	Molecule newMolecules[numOfMolec];
-
-    //need a deep copy of molecule pattern incase it is modified.
-    for (int i = 0; i < moleculePattern.size(); i++)
-    {
-        Atom *atomCopy = new Atom[ moleculePattern[i].numOfAtoms] ;
-        for(int a=0; a <  moleculePattern[i].numOfAtoms ; a++)
-        {
-            atomCopy[a]=  moleculePattern[i].atoms[a];
-        }
-
-        Bond *bondCopy = new Bond[ moleculePattern[i].numOfBonds] ;
-        for(int a=0; a <  moleculePattern[i].numOfBonds ; a++)
-        {
-            bondCopy[a]=  moleculePattern[i].bonds[a];
-        }
-
-        Angle *angleCopy = new Angle[ moleculePattern[i].numOfAngles] ;
-        for(int a=0; a <  moleculePattern[i].numOfAngles ; a++)
-        {
-            angleCopy[a]=  moleculePattern[i].angles[a];
-        }
-
-        Dihedral *dihedCopy = new Dihedral[ moleculePattern[i].numOfDihedrals];
-        for(int a=0; a <  moleculePattern[i].numOfDihedrals ; a++)
-        {
-            dihedCopy[a]=  moleculePattern[i].dihedrals[a];
-        }
-
-        //calculate and add array of Hops to the molecule
-        vector<Hop> calculatedHops;
-        calculatedHops = calculateHops(moleculePattern[i]);
-        int numOfHops = calculatedHops.size();
-        Hop *hopCopy = new Hop[numOfHops];
-        for(int a=0; a < numOfHops; a++)
-        {
-            hopCopy[a] = calculatedHops[a];
-        }
-
-        Molecule molecCopy = Molecule(-1, moleculePattern[i].type, atomCopy, angleCopy, bondCopy, dihedCopy, hopCopy,
-                                    moleculePattern[i].numOfAtoms,
-                                    moleculePattern[i].numOfAngles,
-                                    moleculePattern[i].numOfBonds,
-                                    moleculePattern[i].numOfDihedrals,
-                                    numOfHops);
-		newMolecules[i] = molecCopy;
-
-    }
-
-	//Assign/calculate the appropiate x,y,z positions to the molecules.
-	//buildMoleculeInSpace(newMolecules, numOfMolec);
-	buildMoleculeXYZ(newMolecules, numOfMolec);
-
-
-    for (int i = 0; i < numOfMolec; i++)
-    {
-        if(i == 0)
-        {
-            newMolecules[i].id = startingID;
-        }
-        else
-        {
-	       newMolecules[i].id = newMolecules[i-1].id + newMolecules[i-1].numOfAtoms;
-        }
-    }
-
-    for (int j = 0; j < numOfMolec; j++)
-    {
-        Molecule newMolecule = newMolecules[j];
-        //map unique IDs to atoms within structs based on startingID
-        for(int i = 0; i < newMolecules[j].numOfAtoms; i++)
-        {
-            int atomID = newMolecule.atoms[i].id - 1;
-      	    newMolecule.atoms[i].id = atomID + startingID;
-        }
-        for (int i = 0; i < newMolecule.numOfBonds; i++)
-        {
-            int atom1ID = newMolecule.bonds[i].atom1 - 1;
-            int atom2ID = newMolecule.bonds[i].atom2 - 1;
-
-			newMolecule.bonds[i].atom1 = atom1ID + startingID;
-            newMolecule.bonds[i].atom2 = atom2ID + startingID;
-        }
-        for (int i = 0; i < newMolecule.numOfAngles; i++)
-        {
-            int atom1ID = newMolecule.angles[i].atom1 - 1;
-            int atom2ID = newMolecule.angles[i].atom2 - 1;
-
-			newMolecule.angles[i].atom1 = atom1ID + startingID;
-            newMolecule.angles[i].atom2 = atom2ID + startingID;
-        }
-        for (int i = 0; i < newMolecule.numOfDihedrals; i++)
-        {
-            int atom1ID = newMolecule.dihedrals[i].atom1 - 1;
-            int atom2ID = newMolecule.dihedrals[i].atom2 - 1;
-
-			newMolecule.dihedrals[i].atom1 = atom1ID + startingID;
-            newMolecule.dihedrals[i].atom2 = atom2ID + startingID;
-        }
-        for (int i = 0; i < newMolecule.numOfHops; i++)
-        {
-            int atom1ID = newMolecule.hops[i].atom1 - 1;
-            int atom2ID = newMolecule.hops[i].atom2 - 1;
-
-	        newMolecule.hops[i].atom1 = atom1ID + startingID;
-            newMolecule.hops[i].atom2 = atom2ID + startingID;
-        }
-    }
-
-    return vector<Molecule>(newMolecules,newMolecules+sizeof(newMolecules)/sizeof(Molecule) );
-}
 
 
 // ============================================================================
@@ -2529,82 +1360,76 @@ Real SBScanner::getEqAngle(string endpoint1, string middleAtom, string endpoint2
 // ======================= Logging Functions ==================================
 // ============================================================================
 
-void writeToLog(string text,int stamp)
-{
-    string filename = "OutputLog";
+void writeToLog(string text,int stamp) {
+  string filename = "OutputLog";
 	ofstream logFile;
 	logFile.open(filename.c_str(),ios::out|ios::app);
 
 	string hash ="";
 	time_t current_time;
-    struct tm * time_info;
-    char timeString[9];  // space for "HH:MM:SS\0"
+  struct tm * time_info;
+  char timeString[9];  // space for "HH:MM:SS\0"
 
-	switch(stamp)
-    {
-        case START:
-            //The start of a new simulation
-            logFile << "\n\n\n\n\n\n" << endl;
-            logFile << "======================================================================"<<endl;
-            logFile << "                       Starting Simulation: ";
-            time(&current_time);
-            time_info = localtime(&current_time);
-            strftime(timeString, sizeof(timeString), "%H:%M:%S", time_info);
-            logFile << timeString << endl;
-            logFile << "----------------------------------------------------------------------"<<endl;
-            break;
+	switch(stamp) {
+  	case START:
+    	//The start of a new simulation
+      logFile << "\n\n\n\n\n\n" << endl;
+      logFile << "======================================================================"<<endl;
+      logFile << "                       Starting Simulation: ";
+      time(&current_time);
+      time_info = localtime(&current_time);
+      strftime(timeString, sizeof(timeString), "%H:%M:%S", time_info);
+      logFile << timeString << endl;
+      logFile << "----------------------------------------------------------------------"<<endl;
+      break;
 		case END:
-            //The end of a running simulation
-            logFile << "----------------------------------------------------------------------"<<endl;
-            logFile << "                       Ending Simulation: ";
-            time(&current_time);
-            time_info = localtime(&current_time);
-            strftime(timeString, sizeof(timeString), "%H:%M:%S", time_info);
-            logFile << timeString << endl;
-            logFile << "======================================================================"<<endl;
-            break;
-        case OPLS:
-            //OPLS error
-            logFile << "--OPLS: ";
-            break;
-        case Z_MATRIX:
-            //Zmatrix error
-            logFile << "--Z_Matrix: ";
-            break;
-        case GEOM:
-            //GEOM error Geometric
-            logFile << "--GEOM: ";
-            break;
-        default:
-            logFile << "";
-            break;
+      //The end of a running simulation
+      logFile << "----------------------------------------------------------------------"<<endl;
+      logFile << "                       Ending Simulation: ";
+      time(&current_time);
+      time_info = localtime(&current_time);
+      strftime(timeString, sizeof(timeString), "%H:%M:%S", time_info);
+      logFile << timeString << endl;
+      logFile << "======================================================================"<<endl;
+      break;
+    case OPLS:
+      //OPLS error
+      logFile << "--OPLS: ";
+      break;
+    case Z_MATRIX:
+      //Zmatrix error
+      logFile << "--Z_Matrix: ";
+      break;
+    case GEOM:
+      //GEOM error Geometric
+      logFile << "--GEOM: ";
+      break;
+    default:
+      logFile << "";
+      break;
 	 }
 	 logFile << text << endl;
 	 logFile.close();
 }
 
-void writeToLog(stringstream& ss, int stamp)
-{
-    writeToLog(ss.str(),stamp);
+void writeToLog(stringstream& ss, int stamp) {
+	writeToLog(ss.str(),stamp);
 	ss.str(""); // clears the string steam...
 	ss.clear();
 }
 
-static inline std::string& rtrim(std::string& s)
-{
-    s.erase(s.find_last_not_of(TRIMMED_CHARS) + 1);
-    return s;
+static inline std::string& rtrim(std::string& s) {
+	s.erase(s.find_last_not_of(TRIMMED_CHARS) + 1);
+  return s;
 }
 
-static inline std::string& ltrim(std::string& s)
-{
-    std::size_t len = s.find_first_not_of(TRIMMED_CHARS);
-    if (len != std::string::npos)
-        s.erase(0, len);
-    return s;
+static inline std::string& ltrim(std::string& s) {
+  std::size_t len = s.find_first_not_of(TRIMMED_CHARS);
+  if (len != std::string::npos)
+    s.erase(0, len);
+  return s;
 }
 
-static inline std::string& trim(std::string& s)
-{
-    return ltrim(rtrim(s));
+static inline std::string& trim(std::string& s) {
+  return ltrim(rtrim(s));
 }
